requirements.txt
langchain
openai
python-dotenv
langchain-openai
chardet
langchain-community
langchain-text-splitters
langchain-core